Extract SettingsPageContent to separate file
- Extracted monolithic `SettingsPageContent` from `src/app/settings/page.tsx` to `src/app/settings/SettingsPageContent.tsx`.
- Updated `src/app/settings/page.tsx` to wrap `SettingsPageContent` in a `Suspense` boundary.
- Cleaned up unused imports in `src/app/settings/page.tsx`.
- Verified functionality with Playwright script (manual onboarding bypass).
